Safari-611.1.10/LayoutTests/plugins/document-open-expected.txt

CONSOLE MESSAGE: PLUGIN: DOCUMENT OPEN SUCCESS
